Addons signing
A disaster of manual cherry picking.

Due to a truly bizarre issue with mozsvc 0.8 causing deeply funky issues
with M2Crypto's parameter marshalling, this is an ungodly mash up of the
new rebase that this release was intended to be and the versions currently in
production.

Some tidying of code done but the restructuring of trunion.crypto will have
to happen in a follow up release.

Apps and addons view testing works now, though it's a pretty example of copy
and paste coding.

OAuth support had to be reverted because of the weird M2Crypto issues.